Fixed the distance calculation again, was calculating distances straight from latitude and longitude degrees instead of x and y coordinates.

+++ src/openpaths/OpenPath.java
@@ -39,7 +39,7 @@ public class OpenPath {
      Physical Distance:
      - Finds the physical distance between two locations
      */
-    public float findDistanceBetween(OpenPath secondLocation) {
+    public double findDistanceBetween(OpenPath secondLocation) {
         return vectorPoint.findDistance(secondLocation.getVectorPoint());
     }
